+++ lib/generators/pavlov/templates/backend/backend.rb
@@ -1,3 +1,3 @@
-class Backend
+class Backend < Pavlov::Backend
   # This class is a placeholder that will be expanded by a later feature branch
 end