+++ lib/generators/pavlov/templates/backend/query.rb
@@ -1,3 +1,0 @@
-class Query
-  include Pavlov::Operation
-end