+++ lib/pavlov.rb
@@ -10,32 +10,44 @@ module Pavlov
   end
 
   def self.command command_name, *args
-    class_name = "Commands::"+string_to_classname(command_name)
-    klass = get_class_by_string(class_name)
+    klass = class_for_command(command_name)
     klass.new(*args).call
   end
 
-  def self.interactor command_name, *args
-    class_name = "Interactors::"+string_to_classname(command_name)
-    klass = get_class_by_string class_name
+  def self.interactor interactor_name, *args
+    klass = class_for_interactor(interactor_name)
     klass.new(*args).call
   end
 
-  def self.query command_name, *args
-    class_name = "Queries::"+string_to_classname(command_name)
-    klass = get_class_by_string class_name
+  def self.query query_name, *args
+    klass = class_for_query(query_name)
     klass.new(*args).call
+  end
+
+  private
+
+  def self.class_for_command command_name
+    class_name = 'Commands::' + string_to_classname(command_name)
+    get_class_by_string(class_name)
+  end
+
+  def self.class_for_interactor interactor_name
+    class_name = 'Interactors::' + string_to_classname(interactor_name)
+    get_class_by_string(class_name)
+  end
+
+  def self.class_for_query query_name
+    class_name = 'Queries::' + string_to_classname(query_name)
+    get_class_by_string(class_name)
   end
 end
 
 require_relative 'pavlov/engine' if defined?(Rails)
 require_relative 'pavlov/helpers'
 require_relative 'pavlov/access_denied'
-require_relative 'pavlov/validation_error'
-require_relative 'pavlov/validations'
+require_relative 'pavlov/backend'
 require_relative 'pavlov/operation'
 require_relative 'pavlov/command'
 require_relative 'pavlov/query'
 require_relative 'pavlov/interactor'
 require_relative 'pavlov/version'
-require_relative 'pavlov/alpha_compatibility'