+++ lib/pavlov/concern.rb
@@ -1,0 +1,54 @@
+# Copyright (c) 2005-2011 David Heinemeier Hansson
+#           (c) 2013 Factlink, Inc
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+# fork from active_support 3.2.14 : ActiveSuport::Concern
+# Removed deprecation warning, and don't include InstanceMethods
+#
+# Including InstanceMethods conflicts with Virtus (which also uses InstanceMethods)
+
+module Pavlov
+  module Concern
+    def self.extended(base)
+      base.instance_variable_set('@_dependencies', [])
+    end
+
+    def append_features(base)
+      if base.instance_variable_defined?('@_dependencies')
+        base.instance_variable_get('@_dependencies') << self
+        return false
+      else
+        return false if base < self
+        @_dependencies.each { |dep| base.send(:include, dep) }
+        super
+        base.extend const_get('ClassMethods') if const_defined?('ClassMethods')
+        base.class_eval(&@_included_block) if instance_variable_defined?('@_included_block')
+      end
+    end
+
+    def included(base = nil, &block)
+      if base.nil?
+        @_included_block = block
+      else
+        super
+      end
+    end
+  end
+end