+++ lib/pavlov/helpers.rb
@@ -1,18 +1,15 @@
 module Pavlov
   module Helpers
-    def interactor name, hash
-      hash = add_pavlov_options hash
-      Pavlov.interactor name, hash
+    def interactor name, hash = {}
+      Pavlov.interactor name, with_pavlov_options(hash)
     end
 
-    def query name, hash
-      hash = add_pavlov_options hash
-      Pavlov.query name, hash
+    def query name, hash = {}
+      Pavlov.query name, with_pavlov_options(hash)
     end
 
-    def command name, hash
-      hash = add_pavlov_options hash
-      Pavlov.command name, hash
+    def command name, hash = {}
+      Pavlov.command name, with_pavlov_options(hash)
     end
 
     def pavlov_options
@@ -20,10 +17,9 @@ module Pavlov
     end
 
     private
-    def add_pavlov_options hash
-      if pavlov_options != {}
-        hash ||= {}
 
+    def with_pavlov_options hash
+      if pavlov_options != {}
         if hash.has_key? 'pavlov_options'
           hash[:pavlov_options] = pavlov_options.merge(hash[:pavlov_options])
         else