+++ lib/pavlov/operation_finder.rb
@@ -1,0 +1,32 @@
+module Pavlov
+  class OperationFinder
+    attr_reader :namespace, :backend
+
+    def initialize(namespace, backend, autocall = false)
+      @namespace = namespace
+      @backend   = backend
+      @autocall  = autocall
+    end
+
+    def method_missing(name, attributes = {})
+      klass    = _find_klass(name)
+      instance = klass.new(attributes.merge(backend: backend))
+
+      if @autocall
+        instance.call
+      else
+        instance
+      end
+    end
+
+    def respond_to_missing?(name, include_private = false)
+      _find_klass(name)
+    end
+
+    private
+
+    def _find_klass(name)
+      namespace.constantize.const_get(name.to_s.camelize)
+    end
+  end
+end