+++ lib/pavlov/query.rb
@@ -1,9 +1,13 @@
-require 'active_support/concern'
+require_relative 'concern'
 require_relative 'operation'
 
 module Pavlov
   module Query
-    extend ActiveSupport::Concern
+    extend Pavlov::Concern
     include Pavlov::Operation
+
+    def authorized?
+      true
+    end
   end
 end