+++ script/ci.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+set -e
+
+rake spec
+bundle exec rubocop