+++ spec/pavlov/operation_spec.rb
@@ -9,30 +9,45 @@ describe Pavlov::Operation do
         include Pavlov::Operation
         attribute :validation_check,    Object
         attribute :authorization_check, Object
-        def valid?; validation_check.call; end
-        def authorized?; authorization_check.call; end
-        def execute; end
+
+        def validate
+          validation_check.call
+        end
+
+        def authorized?
+          authorization_check.call
+        end
+
+        def execute
+        end
       end
     end
 
     let(:validation_check)    { double(call: true) }
     let(:authorization_check) { double(call: true) }
-    let(:operation) { dummy_class.new(validation_check: validation_check,
-                                      authorization_check: authorization_check) }
+    let(:operation) do
+      dummy_class.new validation_check: validation_check,
+                      authorization_check: authorization_check
+    end
 
-    it "calls validate if it exists" do
+    it 'calls validate if it exists' do
       operation.call
       expect(validation_check).to have_received(:call).once
     end
 
-    it "calls check_authority" do
+    it 'calls check_authority' do
       operation.call
       expect(authorization_check).to have_received(:call).once
     end
 
     it 'raises ValidationError when not valid' do
-      validation_check.stub(call: false)
+      validation_check.stub(:call).and_raise(Pavlov::ValidationError)
       expect { operation.call }.to raise_error(Pavlov::ValidationError)
+    end
+
+    it 'raises ValidationError when not valid' do
+      authorization_check.stub(call: false)
+      expect { operation.call }.to raise_error(Pavlov::AccessDenied)
     end
   end
 
@@ -43,17 +58,17 @@ describe Pavlov::Operation do
           include Pavlov::Operation
           attribute :title, String
         end
-        expect(dummy_class.new(title: 'Title').validate).to be_true
+        expect(dummy_class.new(title: 'Title').valid?).to be_true
       end
 
       it 'passes when given a default' do
         dummy_class = Class.new do
           include Pavlov::Operation
-          attribute :title,   String, default: "A title"
-          attribute :date,    Time,   default: lambda { Time.now }
+          attribute :title,   String, default: 'A title'
+          attribute :date,    Time,   default: -> { Time.now }
           attribute :visible, String, default: nil
         end
-        expect(dummy_class.new.validate).to be_true
+        expect(dummy_class.new.valid?).to be_true
       end
 
       it 'fails when not given a value' do
@@ -61,7 +76,7 @@ describe Pavlov::Operation do
           include Pavlov::Operation
           attribute :title, String
         end
-        expect(dummy_class.new.validate).to be_false
+        expect(dummy_class.new.valid?).to be_false
       end
     end
   end
@@ -71,30 +86,34 @@ describe Pavlov::Operation do
       stub_const 'Pavlov::AccessDenied', StandardError
     end
 
-    it "raises no error when .authorized? returns true" do
+    it 'raises no error when .authorized? returns true' do
       dummy_class = Class.new do
         include Pavlov::Operation
         def authorized?
           true
         end
-        def execute; end
+
+        def execute
+        end
       end
-      expect {
+      expect do
         dummy_class.new.call
-      }.to_not raise_error
+      end.to_not raise_error
     end
 
-    it "raises an error when .authorized? returns false" do
+    it 'raises an error when .authorized? returns false' do
       dummy_class = Class.new do
         include Pavlov::Operation
         def authorized?
           false
         end
-        def execute; end
+
+        def execute
+        end
       end
-      expect {
+      expect do
         dummy_class.new.call
-      }.to raise_error Pavlov::AccessDenied
+      end.to raise_error Pavlov::AccessDenied
     end
   end
 end